+++ tests/asio_repe/CMakeLists.txt
@@ -1,0 +1,24 @@
+project(asio_repe)
+
+FetchContent_Declare(
+    asio
+    GIT_REPOSITORY https://github.com/chriskohlhoff/asio.git
+    GIT_TAG master
+    GIT_SHALLOW TRUE
+)
+FetchContent_GetProperties(asio)
+if(NOT asio_POPULATED)
+    FetchContent_Populate(asio)
+endif()
+
+add_subdirectory(server)
+add_subdirectory(client)
+
+add_executable(${PROJECT_NAME} ${PROJECT_NAME}.cpp)
+
+target_include_directories(${PROJECT_NAME} PRIVATE include ${asio_SOURCE_DIR}/asio/include ${CMAKE_CURRENT_SOURCE_DIR}/../include)
+target_link_libraries(${PROJECT_NAME} PRIVATE glz_test_common)
+
+add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
+
+target_code_coverage(${PROJECT_NAME} AUTO ALL)
