+++ tests/asio_repe/client/CMakeLists.txt
@@ -1,0 +1,10 @@
+project(repe_client)
+
+add_executable(${PROJECT_NAME} ${PROJECT_NAME}.cpp)
+
+target_include_directories(${PROJECT_NAME} PRIVATE include ${asio_SOURCE_DIR}/asio/include)
+target_link_libraries(${PROJECT_NAME} PRIVATE glz_test_common)
+
+add_test(NAME ${PROJECT_NAME} COMMAND ${PROJECT_NAME})
+
+target_code_coverage(${PROJECT_NAME} AUTO ALL)
