+++ tests/asio_repe/server/repe_server.cpp
@@ -1,0 +1,19 @@
+// Glaze Library
+// For the license information refer to glaze.hpp
+
+#ifndef BOOST_UT_DISABLE_MODULE
+#define BOOST_UT_DISABLE_MODULE
+#endif
+
+#include "boost/ut.hpp"
+
+using namespace boost::ut;
+
+#include "glaze/glaze.hpp"
+#include "glaze/rpc/repe.hpp"
+
+int main()
+{
+   const auto result = boost::ut::cfg<>.run({.report_errors = true});
+   return result;
+}
